requirements.txt
streamlit
bitvavo-python